✨ feat(game): add DELETE /game/:id it deletes also related players, as it is relationnal

+++ back_end/src/modules/game/game.service.ts
@@ -172,7 +172,14 @@ export class GameService {
 
   //TODO: remove en cascade : il faut aussi remove les players correspondant
   async remove(uuid: string) {
-    const game = await this.findOne(uuid);
+    const game = await this.game_repo.findOne(uuid, { relations: ['players'] });
+    if (!game) throw new NotFoundException('game not found');
+
+    for (let i = 0; i < game.players.length; i++) {
+      const player = await this.player_repo.findOne(game.players[i].id);
+      if (player) await this.player_repo.remove(player);
+    }
+
     return await this.game_repo.remove(game);
   }
